Users/ASUS/Desktop/SGP-2025/requirements_flask.txt
Flask==2.3.3
Werkzeug==2.3.7
torch==2.7.1+cpu
torchvision==0.22.1+cpu
torchaudio==2.7.1+cpu
transformers==4.35.2
scikit-learn==1.3.2
numpy==1.24.3
PyPDF2==3.0.1
PyMuPDF==1.26.3
Pillow==10.1.0
pytesseract==0.3.10
requests==2.32.4
python-dotenv==1.0.0 